requirements.txt
# ── Microsoft Agent Framework (MAF) ──────────────────────────
agent-framework @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python
agent-framework-foundry-local @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/foundry_local
agent-framework-orchestrations @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/orchestrations

# ── MCP ──────────────────────────────────────────────────────
fastmcp==3.0.0

# ── Telemetry ────────────────────────────────────────────────
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc

# ── Configuration ────────────────────────────────────────────
python-dotenv

# ── Testing ──────────────────────────────────────────────────
pytest
pytest-asyncio
